Split helper classes from main example widget.
Move the `WebView` and related `WebViewController` classes from the
main.dart into a separate web_view.dart file.
