+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.0.1
+## 2.0.13
 
 * Extract Android implementation from `webview_flutter`.
