+++ packages/webview_flutter/webview_flutter_android/example/pubspec.yaml
@@ -3,7 +3,7 @@ description: Demonstrates how to use the webview_flutter_android plugin.
 publish_to: none
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
